+++ src/api/controladorprecios/app/Http/Controllers/EquivalenciaController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Contractors\IMapper;
 use App\Contractors\Services\IEquivalenciasService;
 use App\DTOs\EquivalenciaDTO;
+use App\Enums\AccionsEnums;
 use App\Mappers\EquivalenciaMapper;
 use App\Services\EquivalenciasService;
 use App\Services\ServicesContainer;
@@ -23,6 +24,7 @@ class EquivalenciaController extends BaseController
 
     public function addEquivalencia(Request $request){
         try {
+            if(!$this->HasAccionsPermit([AccionsEnums::ADD_PERMIT,AccionsEnums::ADD_EQUIVALENCIA_PERMIT])) return new Response($this->stdResponse(false,true,'401 Unauthorized'),401);
             $jsonParsed= $this->validateJsonContent($request);
             if($jsonParsed instanceof Response) return $jsonParsed;
             $dto = $this->mapper->reverse($jsonParsed);
@@ -35,6 +37,7 @@ class EquivalenciaController extends BaseController
 
     public function addEquivalencias(Request $request,$id){
         try {
+            if(!$this->HasAccionsPermit([AccionsEnums::ADD_PERMIT,AccionsEnums::ADD_EQUIVALENCIA_PERMIT])) return new Response($this->stdResponse(false,true,'401 Unauthorized'),401);
             $jsonParsed= $this->validateJsonContent($request);
             if($jsonParsed instanceof Response) return $jsonParsed;
             if(!is_array($jsonParsed)) new Response($this->stdResponse(false,true,"invalid content"),400);
@@ -51,6 +54,7 @@ class EquivalenciaController extends BaseController
 
     public function deleteEquivalencia($publicId){
         try {
+            if(!$this->HasAccionsPermit([AccionsEnums::DELETE_PERMIT,AccionsEnums::DELETE_EQUIVALENCIA_PERMIT])) return new Response($this->stdResponse(false,true,'401 Unauthorized'),401);
             $this->service->deleteEquivalencia($publicId);
             return new Response($this->stdResponse());
         } catch (\Throwable $th) {
